raiderioWatch/pyproject.toml
[tool.poetry]
name = "raideriowatch"
version = "0.1.0"
description = "raider io data pipeline service"
authors = ["tom pianka <tomyuval8@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.34.31"
botocore = "^1.34.31"
docker = "^7.0.0"
logging = "^0.4.9.6"
asyncio = "^3.4.3"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
